requirements.txt
appdirs==1.4.3
Babel==2.5.3
Django==2.0.4
django-attachments==1.3
django-bootstrap-breadcrumbs==0.9.1
django-bootstrap3==9.1.0
django-bootstrap3-datetimepicker-2==2.5.0
django-ckeditor==5.4.0
django-filter==1.1.0
django-js-asset==1.0.0
django-localflavor==2.0
django-phonenumber-field==2.0.0
django-robots==3.0
django-smart-selects==1.5.4
django-widget-tweaks==1.4.2
flake8==3.5.0
image==1.5.20
mccabe==0.6.1
olefile==0.45.1
packaging==17.1
phonenumberslite==8.9.2
Pillow==5.1.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pycodestyle==2.3.1
pyflakes==1.6.0
pyparsing==2.2.0
pytz==2018.3
six==1.11.0
